Resolve SeqType based on a compiler flag
Relying on the classpath didn't work as it was sometime preferred over
the source path. We need to knows which standard library we are
compiling against even before symbols are entered (e.g. during
desugaring).

This flag will go away once we drop support for the 2.12 standard
library. But for now, we need it to compile the 2.13 standard library.
