+++ compiler/src/dotty/tools/dotc/ast/Desugar.scala
@@ -486,7 +486,7 @@ object desugar {
 
       // TODO When the Scala library is updated to 2.13.x add the override keyword to this generated method.
       // (because Product.scala was updated)
-      def productElemNameMethod = {
+      def productElemNameMeth = {
         val methodParam = makeSyntheticParameter(tpt = scalaDot(tpnme.Int))
         val paramRef = Ident(methodParam.name)
 
@@ -499,11 +499,11 @@ object desugar {
         } :+ defaultCase
         val body = Match(paramRef, patternMatchCases)
         DefDef(nme.productElementName, Nil, List(List(methodParam)), javaDotLangDot(tpnme.String), body)
-          .withFlags(Synthetic)
+          .withFlags(if (ctx.settings.YnewLibrary.value) Override | Synthetic else Synthetic)
       }
 
       if (isCaseClass)
-        productElemNameMethod :: copyMeths ::: enumTagMeths ::: productElemMeths
+        productElemNameMeth :: copyMeths ::: enumTagMeths ::: productElemMeths
       else Nil
     }
